requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
opencv-python-headless
seaborn
Pillow
albumentations
tensorflow-datasets
flask


